General fixes branch (#27)
* changed the Log Out button color

* Removed redirect from Login page. Changed the alerts for toasts.

* Dashboard: changed the loged out message. Store
-Changed  headings/titles sizes. Substituted the page's tab icon.

* Changed all alerts to the final ones. Added avatar image to posts, when the user has an avatar.

* changed spinner color posts

* Loaders/Spinners added

* Added Spinners for loading state. Added stars colors

* Delete address feature incomplete

* added functionallity to DELETE adress button in the profile's page.

* Design:
-Added shadows to H1  Headings √
-Changed Price color in the Single Products page √
Products-changed the text heading on the products page (Alert) for normal headings and a  div box.
Store - Removed the margins swapping the Container for a div
Login and SignUp - centralized and created a divbox in the form's background
Created Footer component and added global css and body elemnt in the layout.tsx to bring it to the pages bottom.